backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlmodel==0.0.16
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.1

requests==2.31.0
